added test_quantifiers.py and changed previous tests imports
