+++ tests/html/test_tabular.py
@@ -1,8 +1,8 @@
 from pytest import fixture
 from lxml import etree
-from libextract.html.tabular import children_counter, \
-        get_node_counter_pairs, node_counter_argmax, \
-        sort_best_pairs, weighted_score, filter_tags
+from libextract.html.tabular import get_node_counter_pairs, \
+        node_counter_argmax, sort_best_pairs, weighted_score, \
+        filter_tags
 
 
 @fixture
@@ -19,14 +19,6 @@ def article(etree):
 def sorted_pairs(pairs):
     return sort_best_pairs(node_counter_argmax(pairs),
                            top=1)
-
-
-def test_children_counter(etree):
-    article = etree.xpath('//body/article')[0]
-    counter = children_counter(article)
-
-    assert len(counter) == 1
-    assert counter['div'] == 9
 
 
 def test_get_node_counter_pairs(pairs):
