+++ tests/test_quantifiers.py
@@ -1,0 +1,15 @@
+from libextract.quantifiers import text_length, children_counter
+
+
+def test_text_length(etree):
+    res = etree.xpath('//body/article/div')
+    for node in res:
+        assert text_length(node) == 4
+    assert res
+
+def test_children_counter(etree):
+    article = etree.xpath('//body/article')[0]
+    counter = children_counter(article)
+
+    assert len(counter) == 1
+    assert counter['div'] == 9
